fix tests hopefully
